+++ rl/da.go
@@ -6,7 +6,6 @@ package rl
 
 import (
 	"github.com/emer/axon/axon"
-	"github.com/emer/emergent/emer"
 	"github.com/goki/ki/kit"
 )
 
@@ -19,61 +18,13 @@ type DALayer interface {
 	SetDA(da float32)
 }
 
-// SendDA is a list of layers to send dopamine to
-type SendDA emer.LayNames
-
-// SendDA sends dopamine to list of layers
-func (sd *SendDA) SendDA(net emer.Network, da float32) {
-	for _, lnm := range *sd {
-		ml, ok := net.LayerByName(lnm).(DALayer)
-		if ok {
-			ml.SetDA(da)
-		}
-	}
-}
-
-// Validate ensures that LayNames layers are valid.
-// ctxt is string for error message to provide context.
-func (sd *SendDA) Validate(net emer.Network, ctxt string) error {
-	ln := (*emer.LayNames)(sd)
-	return ln.Validate(net, ctxt)
-}
-
-// Add adds given layer name(s) to list
-func (sd *SendDA) Add(laynm ...string) {
-	*sd = append(*sd, laynm...)
-}
-
-// AddOne adds one layer name to list -- python version -- doesn't support varargs
-func (sd *SendDA) AddOne(laynm string) {
-	*sd = append(*sd, laynm)
-}
-
-// AddAllBut adds all layers in network except those in exlude list
-func (sd *SendDA) AddAllBut(net emer.Network, excl ...string) {
-	ln := (*emer.LayNames)(sd)
-	ln.AddAllBut(net, excl...)
-}
-
-// Layers that use SendDA should include a Validate check in Build as follows:
-
-// Build constructs the layer state, including calling Build on the projections.
-// func (ly *DaSrcLayer) Build() error {
-// 	err := ly.Layer.Build()
-// 	if err != nil {
-// 		return err
-// 	}
-// 	err = ly.SendDA.Validate(ly.Network, ly.Name()+" SendTo list")
-// 	return err
-// }
-
 //////////////////////////////////////////////////////////////////////////////////////
 //  ClampDaLayer
 
 // ClampDaLayer is an Input layer that just sends its activity as the dopamine signal
 type ClampDaLayer struct {
 	Layer
-	SendDA SendDA `desc:"list of layers to send dopamine to"`
+	SendDA axon.SendDA `desc:"list of layers to send dopamine to"`
 }
 
 var KiT_ClampDaLayer = kit.Types.AddType(&ClampDaLayer{}, LayerProps)
