+++ rl/layer.go
@@ -5,30 +5,22 @@
 package rl
 
 import (
-	"fmt"
-
 	"github.com/emer/axon/axon"
 	"github.com/emer/axon/deep"
 	"github.com/emer/emergent/emer"
 	"github.com/goki/ki/ki"
 	"github.com/goki/ki/kit"
-	"github.com/goki/mat32"
 )
 
 // Layer is the base layer type for RL framework.
 // Adds a dopamine variable to base Axon layer type.
 type Layer struct {
 	axon.Layer
-	DA float32 `inactive:"+" desc:"dopamine value for this layer"`
 }
 
 var KiT_Layer = kit.Types.AddType(&Layer{}, LayerProps)
 
 // DALayer interface:
-
-func (ly *Layer) GetDA() float32   { return ly.DA }
-func (ly *Layer) SetDA(da float32) { ly.DA = da }
-
 func (ly *Layer) Defaults() {
 	ly.Layer.Defaults()
 	ly.Typ = RL
@@ -36,50 +28,6 @@ func (ly *Layer) Defaults() {
 
 func (ly *Layer) Class() string {
 	return "RL " + ly.Cls
-}
-
-// UnitVarIdx returns the index of given variable within the Neuron,
-// according to UnitVarNames() list (using a map to lookup index),
-// or -1 and error message if not found.
-func (ly *Layer) UnitVarIdx(varNm string) (int, error) {
-	vidx, err := ly.Layer.UnitVarIdx(varNm)
-	if err == nil {
-		return vidx, err
-	}
-	if varNm != "DA" {
-		return -1, fmt.Errorf("rl.Layer: variable named: %s not found", varNm)
-	}
-	nn := ly.Layer.UnitVarNum()
-	return nn, nil
-}
-
-// UnitVal1D returns value of given variable index on given unit, using 1-dimensional index.
-// returns NaN on invalid index.
-// This is the core unit var access method used by other methods,
-// so it is the only one that needs to be updated for derived layer types.
-func (ly *Layer) UnitVal1D(varIdx int, idx int) float32 {
-	nn := ly.Layer.UnitVarNum()
-	if varIdx < 0 || varIdx > nn { // nn = DA
-		return mat32.NaN()
-	}
-	if varIdx < nn {
-		return ly.Layer.UnitVal1D(varIdx, idx)
-	}
-	if idx < 0 || idx >= len(ly.Neurons) {
-		return mat32.NaN()
-	}
-	return ly.DA
-}
-
-// UnitVarNum returns the number of Neuron-level variables
-// for this layer.  This is needed for extending indexes in derived types.
-func (ly *Layer) UnitVarNum() int {
-	return ly.Layer.UnitVarNum() + 1
-}
-
-func (ly *Layer) InitActs() {
-	ly.Layer.InitActs()
-	ly.DA = 0
 }
 
 //////////////////////////////////////////////////////////////////////////////////////
